webadmin: change allocation policy behavior while creating new disk
When creating a new disk using the UI, the disk allocation policy
changes according to the storage domain type
(Block - preallocation, File - thin),
the problem is that the allocation policy changes each time that
the storage domain changes, even if the user selected/changed
the allocation policy by himself.

Because of bug 1093025 will not be fixed, this patch adds a specific
solution for this case by adding a flag which marks whether the user
customizes the allocation policy field or not.
If the user customized the field - the allocation policy will not
change anymore when changing the storage domain to another.

Change-Id: I3b5744ae097b8dc94f488a7a70595b55638af159
Bug-Url: https://bugzilla.redhat.com/1069269
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>