core: fix TryBackToAllSnapshotsOfVm lease storage domain validation
When having a VM that doesn't
have a lease but has a snapshot that contains
a lease.
If the lease storage domain isn't active then
when trying to perform a 'regular preview' to
the snapshot the operation should fail, but
when trying to perform a 'custom preview'
to the snapshot without the lease,
the operation should succeed.

Currently, 'custom preview' to the snapshot
without the lease fails due to the same
validation error of the 'regular snapshot preview'.
This patch fixes this behavior by skipping this
validation if no lease was selected to the
custom preview.

Change-Id: I93d4378280ea1d0c28c18301f1ec810359f54ecb
Bug-Url: https://bugzilla.redhat.com/1579366
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>