core: process OVF in parallel to spead up engine startup
Motivation
The startup process is parsing all stateless snaptshots of running vms
to compute the current macpool. The processing is mad sequentially which
leads failing jboss boot when the there are a lot of snapshots to
process. We have to prevent that from happening.

Modification
Use parallel stream processing - with 60 snapshots it shows 2x
improvement in time. With bigger numbers it may even be bigger

Change:
pass criteria - Must process 3000 snapshots in less than 1 minute

Change-Id: Iea9dfcea7a9844a6787dac42ae36265317a71075
Bug-Url: https://bugzilla.redhat.com/1579008
Signed-off-by: Roy Golan <rgolan@redhat.com>