searchbackend: Don't join tags on users by default
It is useless to join searches on vdc_users_with_tags and it performs
purely for most of the use-cases

If the word 'tag' or 'vm' for 'Users' doesn't apear in the search
syntax then don't cross join it.

This produces better performing queries for the default UI views or api
searchs for users.

Change-Id: I9b08d3758b7b04ac576f78f59311f99c468c66c0
Bug-Url: https://bugzilla.redhat.com/1578357
Signed-off-by: Roy Golan <rgolan@redhat.com>
Signed-off-by: Martin Perina <mperina@redhat.com>