core: Improve MergeStatusCommand
Currently, to verify whether live merge job succeffully completed,
qemu volume chain is intersected with the top volume and the base
volume, if the result is not a list of length 1 (i.e. the top volume
wasn't deleted), we fail live merge. This logic isn't correct all the
time. There are use cases where qemu chain is different than the
chain in the database, e.g. when snapshot creation or deletion fails.

In this patch, we check that the top volume isn't part of qemu chain.
If it is part of the chain, we fail live merge.

Change-Id: Ib8cc19b6c86001e63f86967cae72442157a7a3a5
Bug-Url: https://bugzilla.redhat.com/1554369
Signed-off-by: Ala Hino <ahino@redhat.com>