packaging: setup: Keep some httpd params on update
In the past, we only touched apache httpd configuration on the first
setup, never on upgrades.

In a recent patch we made engine-setup change it also on upgrades.

This broke configurations where the user manually changed the
configuration since the first setup, in particular in cases the user
wanted to use a 3rd party CA.

This patch adds an attribute to each parameter 'only_if_unconfigured',
which marks it to be updated only if it's a new setup, and sets that one
to True only for the items we want to update on upgrades - currently
only SSLProtocol.

Bug-Url: https://bugzilla.redhat.com/1576377
Change-Id: Ia5c7445d01377d1b1383569c9f6235e32fd5a6ba
Signed-off-by: Yedidyah Bar David <didi@redhat.com>
(cherry picked from commit 27f22ec3801dd1b3d759da11926e49c6ed135f48)