db:fix upgrade issue on NULL values
Script 04_02_0960_default_all_search_engine_fields_to_empty_string.sql
sets various table columns values to empty string instead of NULL.
script
04_03_0180_default_all_search_engine_string_fields_to_not_null.sql
forces those columns to be NOT NULL.
If a system is upgraded and includes the 04_02_0960 script but did not
include the 04_03_0180 script, then later on when this system is
upgraded again and includes this time the 04_03_0180 script it may fail
since meanwhile while the system was running NULL values were inserted
to those columns.

This patch solves this issue by modifying the 04_03_0180 script to run
the content of the 04_02_0960 first.

Change-Id: I9c7462e1dc2e1435b62a0f9079e0413830606156
Signed-off-by: emesika <emesika@redhat.com>