webadmin: Extend "Manage Domain" popup
Screenshot: https://imgur.com/a/VDJMxSU

This patch extends the 'Manage Domain' popup.

Since the same table (iscsi luns list) appears also on managing and
creating new storage domain, and also on creating new direct lun disk
popups, this patch deals with some ui changes on those popups too.

Moreover, some more massive changes are expected to be later on, but not
as a part of this patch.

Change-Id: I866b60fcb1e7f5a154884edf789308683722e8d2
Bug-Url: https://bugzilla.redhat.com/1115607
Signed-off-by: Shani Leviim <sleviim@redhat.com>