engine : External VMs prevent placing host in maintenance mode
External running VMs prevent the host to
placed in maintenance. We should ignore
external VMs when moving hosts to maintenance
based on a newly introduced engine config variable
MaintenanceVdsIgnoreExternalVms. By default the
value is false, so the external VMs are not
ignored when host is moved to maintenance. A
value of true for the engine config variable will
move the host to maintenance even when external VMs
are running.

Change-Id: I9e53da29f4430916738e00945b926340beb8b149
Bug-Url: https://bugzilla.redhat.com/1579303
Signed-off-by: Ravi Nori <rnori@redhat.com>