core: Set vramMultiplier for all x86 Linux systems
vram size is set to basic value of 8 MB for all guest OSes except for RHEL 7.
However that assumes that the user selects RHEL 7 as the guest OS in the VM
configuration.  If another item, such as Linux, is used then vram size may be
insufficient.  When the user installs X.Org in the guest OS, the limited vram
size may result in segmentation fault.

Since SPICE supports only Windows, RHEL 6, and RHEL 7 guest systems and vram is
ignored for Windows and RHEL 6, it's safe to set vram to the RHEL 7 value for
all Linux x86 based systems.  If the RHEL 7 driver is used eventually then vram
size is sufficient and otherwise the value is simply ignored.

For those reasons this patch sets vramMultiplier to 2 for all Linux x86 based
guest systems and not just only for RHEL 7.

Change-Id: I938ad829073e08b16db076c1a86890494809e0c6
Bug-Url: https://bugzilla.redhat.com/1520848
Backport-To: 4.2
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>