core: Fix SR-IOV virtual functions leak
When the VM is going from "Powering up" to "Shutdown" state it leads
to virtual functions leak. The VM creation flow is that upon creation
VM is added to set ResourceManager.asyncRunningVms. The VM resides in
this set until it goes from "Powering up" into "Up" or "Shutdown".

The SR-IOV release is done in ProcessDownVmCommand which is called
upon upon clearing the VM resources but only if the flag
VmAnalyzer.movedToDown is set. For the normal flow (not migration)
this flag is set by VmAnalyzer.clearVm function only in case that the
VM is no longer in the asyncRunningVms set.

The issue is that the clearVm for flow from "Powering up" to "Shutdown"
is called before the function that clears VM from the asyncRunningVms.
This leads to movedToDown not to be set and thus the
ProcessDownVmCommand is never called.

RefreshCapabilities does not help to clear the resource leak because
virtual fucntions are associated with its VM in engine. Even if the
resource is not physically allocated to the VM  the engine thinks
opposite.

Fix the VFs leak by switch call places for clearing VM from
asyncRunningVms and clearVm.

Change-Id: I118be64c705e5eaba71071186a7b2d303c30d888
Bug-Url: https://bugzilla.redhat.com/1586019
Signed-off-by: Ales Musil <amusil@redhat.com>
(cherry picked from commit 2c01a61d92ac870caa5276bafe4a2b91c420a977)