core: fix DB deadlock while commiting a snapshot
A database deadlock occurs when restoring a regular snapshot
without memory and having a later snapshot with memory that
should be removed.

When restoring a snapshot, all later snapshots configuration
should be marked as broken. if a snapshot with memory exist
in the chain and should be removed, another database operation
will take place on the same entity when trying to remove the
snapshot from the base_disk table.

In order to prevent this deadlock, the removal of the memory
disks is done in best effort.
If the memory disks removal will fail, the user can remove them
manually.

Change-Id: I9c033919b219cea4fb1bd0d1480621acd45caa05
Bug-Url: https://bugzilla.redhat.com/1588698
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>
(cherry picked from commit 4c21cd07b1c2352ff3ba3464afc43eea1c09019b)