restapi: Error Getting Vm Console When Vm Suspended
When a VM is suspended, the engine purposely does not allow
obtaining a console connection to it. This resulted in an
obscure 500 Internal Server Error in the API. This patch
makes is so status 409 (CONFLICT) is retuned with a clear
error message.

Bug-Url: https://bugzilla.redhat.com/1591271
Change-Id: I35e677b5710ede81c8431025b508b07de0fe0612
Signed-off-by: Ori_Liel <oliel@redhat.com>
(cherry picked from commit 51c013dedeb3ea99298136abb89c8cf892708dde)