core: support template's disk copies in OVF
Today, the template's disk data in the OVF contains only the original
storage domain ID rather than all the storage domains IDs of its copies.

This patch adds another "StorageId" property for each such a copy, so
that each template's disk can contain one or more "StorageId" fields.

Also, it fixes the validations done in
ImportVmTemplateFromConfigurationCommand to require at least one copy of
the disk and not fail the whole command if the first storage domain in
the image's storageIds list does not exist.

Change-Id: I65428a95d033694d7d11ed907bb6b0497aa986e0
Bug-Url: https://bugzilla.redhat.com/1535603
Signed-off-by: Idan Shaby <ishaby@redhat.com>