core: ansible playbook for measuring an image
QEMU recently added an interesting feature of measuring an image
that returns both the virtual size and the actual size, considering
all its volume-chain, of the image.

This capability is now leveraged for OVA creation. Now that we can
retrieve the actual size of the "collapsed" image in qcow2 format,
we no longer need to create temporary volumes that take a long time
and require us to modify some of the exported entity.

This patch adds an ansible-playbook for this operation to keep
the export-to-ova flow VDSM-agnostic.

Change-Id: I5962444aacc88732b6ffbbd1be7c68ce2793bb01
Bug-Url: https://bugzilla.redhat.com/1526033
Signed-off-by: Arik Hadas <ahadas@redhat.com>