core: extract ova through a loopback device
Simplify the OVA extraction by mounting the disk portion of the
OVA section as a loop device and then converting it with qemu-img
to the destination image.

This allows us not only to remove large portion of custom python
code for copying the data, but also that would make it possible
to convert the data to the raw format later on.

Change-Id: I10450708e646486b61fbb70c515e9d9d9056ba49
Bug-Url: https://bugzilla.redhat.com/1526033
Signed-off-by: Arik Hadas <ahadas@redhat.com>