core: Fix extending of a storage domain on a multipath environment
Commit b47af42 changed the  method convertToStatusList in
ServerConnectionStatusReturn to streams however it used the default
collection to map method which raises an exception if a duplicate key
is mapped, while this worked well on regular environments in multipath
environments multiple statuses return causing the mapping to fail on a
duplicate key.
This patches solves that by using a merge function.

Change-Id: I5bc3c99ab43c429c03c91a5597565d6133ca9a8e
Bug-Url: https://bugzilla.redhat.com/1587961
Bug-Url: https://bugzilla.redhat.com/1594552
Signed-off-by: Tal Nisan <tnisan@redhat.com>