engine: Fail with a clear message if cache dir is missing
Cache dir is expected to exist under /var/cache/ovirt-engine.
If the folder is missing, i.e. removed by the user, the installation
process used to fail without a clear message for the failure reason.

The patch adds a clear message for the host deploy failure, in case of a
failure to create a file under the cache directory.

Change-Id: I8695b62227f25a7d07160519e330ca9b7c5c77f5
Bug-Url: https://bugzilla.redhat.com/1591751
Signed-off-by: Moti Asayag <masayag@redhat.com>