core: Apply MTU on autodefined network
When creating a host network, an associated external network
is created via the AutodefineExternalNetworkCommand.
The MTU of the external network is set to the same value like
the host network.
The external network does not require a smaller MTU than the
host network, because it is not tunneled.

If the external provider does not accept the MTU attribute,
this change will cause Autodefine to fail.

Change-Id: I9972e6f3c3014506efb45958233906b4ba295496
Bug-Url: https://bugzilla.redhat.com/1451342
Signed-off-by: Dominik Holler <dholler@redhat.com>
(cherry picked from commit 1d3fe3eada393fd8d1b5de0a9295f290b66a13d6)