engine: Update libvirtVmXml to consider vmBase.multiQueuesEnabled attribute
VmBase.multiQueuesEnabled will represent the engine should calculate the
optimal number of queues per each vnic.

To keep backward compatibility, queues defined in the vnic's profile
custom properties will override True value.

This patch updates libvirtVmXml to consider vmBase.multiQueuesEnabled
attribute when writing the interface.

Due to a bug, queues defined in the vnic profile properties were written to
the xml twice. Once as the 'queues' and the second time as a 'custom'
value.
The code was fixed so if the 'queues' were added to the xml, they won't be
added as custom value.

Change-Id: Iab4c7f09099db18d734c57d1f4cd0633df60e0e9
Bug-Url:https://bugzilla.redhat.com/1574771
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>