core: fix the correlation logic of cd-roms
When user-aliases are not set (that can happen when the a VM
that was started in lower cluster version is migrated to 4.2
or when centos 7.3/7.4 is used in cluster versiov 4.2), we
need some logic to determine the correlation between the reported
devices and those in the database.

The current logic for CD-ROMs is totally wrong - it relies on the
path inside the spec-params of the device. This patch changes that
logic, relying on having a single 'ordinary' CD-ROM and the non-
ordinary ones (payload/vm-init) point to images in the following
path: /var/run/vdsm/payload.

Change-Id: I1708ac4298b35d87399af652288a761eaae05a8d
Bug-Url: https://bugzilla.redhat.com/1595489
Signed-off-by: Arik Hadas <ahadas@redhat.com>