core: restore original default of ClientInactivityTimeout
Remove increase script of TransferImageClientInactivityTimeoutInSeconds
config value. It was introudced only in master branch, so no need
to handle upgrade. We don't need such large value as we rely
now on idle_time in the image ticket. We probably want to add
a separate config value for the API, as it can have a larger
value since the api user has full control on it.

Change-Id: Ib5d9b9888a95f5f92d001be1829805ec33a7e080
Signed-off-by: Daniel Erez <derez@redhat.com>