engine: Not reporting queues=1 since it is the default.
Due to a libvirt/vdsm bug, if we explicitly set queues=1,
we don't get the attribute back when querying libvirt.
We instead get it when using any other value.

That causes hot-unplug to fail.

Since it causes the OST to fail, this patch is a workarround till the vdsm
patch will be ready.

Change-Id: Ica13325b992fe96367e19344435f2fc811b3d77e
Bug-Url: https://bugzilla.redhat.com/1574771
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>