core: add runtime name to vms
The name of VMs used to reflect their name in libvirt, i.e., showing
what the running instance of the emulator is set with. If the name
of the VM changed, we considered it a next-run property that would
be applied when the emulator process is started with the new name.

However, it was not warmly accepted by our users. They want the
name to be changed instantly in the database. It doesn't matter what
the name in libvirt/qemu is because the name is more a part of the
management layer than of the lower/technical layer that should use
UUIDs as they are unique and constant. I couldn't agree more.

Thus, this patch changes VmStatic#name to be set on every status
of the VM, not only when it's Down. Instead, we now have another
field VmDynamic#runtimeName that stores the name used by qemu. This
field would be used by the frontend to notify users that the VM
runs with a different name than the one they see at the moment.

Change-Id: I7b76fe98e3bc7be0b4e83830744e0fa157d2f21b
Bug-Url: https://bugzilla.redhat.com/1454673
Signed-off-by: Arik Hadas <ahadas@redhat.com>