core: log running vms whose name changes
This patch adds a log that is printed when changing the name of
a running VM, specifying the new name and the one the VM was
started with.

Generally, it is not a good practice to complicate the code for
no good reason and I don't think we should promote users to
rely on vm names when debugging issues (they should use UUIDs)
but as this change doesn't worth the time of arguing about that,
let's just add it for whoever it may be useful for.

Change-Id: If270137140a4fe3bfce008177bda9e93839d5e28
Bug-Url: https://bugzilla.redhat.com/1454673
Signed-off-by: Arik Hadas <ahadas@redhat.com>