engine: poly dhcpv6 + autoconf iface config
A NIC ipv6 boot protocol configuration with both IPV6_AUTOCONF and
DHCPV6C set to 'yes' is not supported by engine, although it is a
legal configuration in linux OSs. Until now engine ignored the
IPV6_AUTOCONF=yes setting and set it to 'no'.

This patch introduces support for this 'poly' ipv6 boot protocol
option.
This patch is a backport from master to 4.2

Change-Id: I57955515b4725db4565d75cb5fba703ef301cea0
Bug-Url: https://bugzilla.redhat.com/1360839
Signed-off-by: Eitan Raviv <eraviv@redhat.com>