core: fix TransferImageStatus for resume from UI
As part of Change-Id: I510ae8d0ab2d2af8ae9b3a4789dfc55aa92824e0,
a validation was added for resuming a transfer that should be
only in paused state. However, when resuming an upload from
the webadmin, we use TransferImageStatusCommand just to fetch [*]
the ImageTrnasfer, i.e. 'getParameters().getUpdates()' is null.
Hence, we get an NPE on 'validate' of TransferImageStatusCommand.

[*] see 'UploadImageHandler -> resume'

Change-Id: I67805fa325de91e40a41f9073175e523253a2eea
Bug-Url: https://bugzilla.redhat.com/1601227
Signed-off-by: Daniel Erez <derez@redhat.com>