core: remove disks filtering from validateStorageDomains() when running a VM
When running a VM, all the storage domains that hold the VM disks
should be activated.

Currently, the status of the storage domain checked only if it
holds read-write thin-provisioned disks, so in case of a VM
with only preallocated / RO disks, there will be no validation
for the storage domains status.

This patch removes the filtering of the read-only and
preallocated disks from the list of the disks that their storage
domain should be validated.

Change-Id: Iae1b9e98e5a00c2f8a6bf0cba566ca15b8657c0a
Bug-Url: https://bugzilla.redhat.com/1602339
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>