engine: infinite loop of setupNetworksTask.isDone()
The following flow causes the infinite loop -

JsonRpcVdsServer.setupNetworks executes the FutureTask.
The execution creates a FutureMap and stores it inside the 'map' data
member of the FutureCallable.

The creation the FutureMap may throw TransportRunTimeException in case of
communication issues with the vdsm.
If the exception is thrown an exception is set on the FutureTask and the
FutureTask is marked as 'done'.

When FutureTask.get will be invoked the exception will be thrown.

Since the 'isDone' method of the FutureTask was overriden to check the
'isDone' of the FutureCallable and not of the FutureTask, it always
returns false although the FutureTask is done.

The reason the 'isDone' method of the FutureCallable returns false is that
it always returns false when the 'map' is null (which is null because of
the original exception).

Note - the exception will be caught and printed when
HostSetupNetworkCommand will invoke 'setupNetworksTask.get.

Change-Id: Ia3ecf377876a1a1cfd508254b477e71715e9f7c3
Bug-Url: https://bugzilla.redhat.com/1597062
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>