webadmin: soundcard enabling value should depend on ppc arch
Soundcard device enabling default value is set according to the VM type
(true for Desktop type, false for Server, HP), but in case of a VM run
on ppc arch cpu, the souncard should always be set to disabled/non-changeable
since not supported.

Up till this fix, the soundcard enabled field was set to non-changeable in
case of ppc arch, but enabling value was chosen only according to VM
type.

Change-Id: Ia17ea7e3ae69e2caed199fd8e203a34c2a48de5c
Bug-Url: https://bugzilla.redhat.com/1571563
Signed-off-by: Sharon Gratch <sgratch@redhat.com>