webadmin: upload image - use exact image size
As part of image upload procedure we create a new disk,
it's size should be defined according to this logic:
- QCOW2 image: use virtual size (from header).
- RAW image: use file size (actual size equals virtual size).

This would ensure a new disk with the exact required size
(so no need for resizing). In the UI, we would still use
the rounded up value - as we display sizes in GiB granularity.

Change-Id: I1af9de0b4cf61b02df39777ad55ca71904d3297e
Bug-Url: https://bugzilla.redhat.com/1608716
Signed-off-by: Daniel Erez <derez@redhat.com>