webadmin: Set CD images for a template with ISO
While creating a new VM from a template with attached ISO CD-ROM,
sometimes there's an NPE on the UI.

This one happens due to the asynchronous query that populates the list
of template's ISO images, which may not return its result on time.
Therefore, trying to stream an empty list causes an NPE on the UI.

This patch calls that query before populating the list and fixes it.

Change-Id: I65bbcab7a18149f752567960e6a9017921bd625c
Bug-Url: https://bugzilla.redhat.com/1607131
Signed-off-by: Shani Leviim <sleviim@redhat.com>