webadmin: Setting the visibility of sourceStorageLabel
Screenshot: https://imgur.com/a/tnmAKhm

This patch sets the visibility of sourceStorageLabel on
DisksAllocationItemView.java.
Since this view common for disk operations (copy/move) and disk's
allocation (when creating a VM from a template), the visibility of
the source storage domain label affects the column's alignment.

Therefore, this patch uses a listener, so the label's availability
determines its visibility.

Change-Id: I267da745c8579ba4388a5277990ce614a8501d02
Bug-Url: https://bugzilla.redhat.com/1609552
Signed-off-by: Shani Leviim <sleviim@redhat.com>