restapi: Fix Disk-Attachment hrefs under Export Domain
Under export storage domain, disk-attachment of a VM do not have hrefs
due to limitation of link-generation infrastructure.

The infrastructure identifies a location in the API-tree by the combination:
entity-type, parent-type.

In this particular case, the combination of disk-attachment under a VM exists
twice - once under root, once under Storage-Domain:

.../api/vms/xxx/diskattachments
.../api/storagedomains/zzz/vms/xxx/diskattachments

This patch adds manual creation of the href to overcome this problem

Bug-Url: https://bugzilla.redhat.com/1374323
Change-Id: I479575a41ad2f3b40d33566f444126c202f7a9cb
Signed-off-by: Ori_Liel <oliel@redhat.com>