scheduler: Prefer hosts where the VM fits to a NUMA node
If a VM without NUMA nodes is run on a host with NUMA,
it is good if the VM can fit in a single NUMA node,
otherwise it will impact its performance.

This patch adds a scheduling unit, so that these hosts
are preferred.

The unit is not enabled by default, because it can create
corner cases that may not be desirable.

For example, in the following setup:

- 9 hosts with 16 GB per NUMA node
- 1 host with 4 GB per NUMA node

When we try to run VMs with 6 GB of memory, the shceduling
unit would prevent them from starting on the host with 4 GB
per NUMA node. No matter how overloaded the other hosts are.
It would use the last host only when all the others are full.

Change-Id: I58fa53a55f2bf7d9caf96c053adb6682dda23bb3
Bug-Url: https://bugzilla.redhat.com/1559694
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>