engine: Network events form vdsm should invoke the correct subscriber
Before this fix, no matter what host sent the event, the same default
subscriber was invoked. So in case of network event from the host,
getCapablities was always invoked on the same host. No matter what host
actually sent the event.

The reason was that the 'unique_id' part of the 'subscriptionId' was
set to 'no_id' for the subscribers of all the hosts.
When the 'unique_id' is specified, the subscriber is registered with the
'unique_id' as the key. Only one subscriber is alowed per 'unique_id'.
Since all the 'HostConnectionRefresher' subscribers had the same
'unique_id', only the last one was actually registered by
'SubscriptionMatcher.add'.
Only when the 'unique_id' is not specified (== has '*' value), the other
parts of the 'subscriptionId' are registered by 'SubscriptionMatcher.add'.

So before thise fix, all the other parts of the 'subscriptionId' were
ignored by 'SubscriptionMatcher.add'.

Change-Id: Iaa73a68619d7c5787b71d59158db1a5b79755619
Bug-Url: https://bugzilla.redhat.com/1590109
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>