core: introduce VmLeasesReadyFilterPolicyUnit
After activating a storage domain, it takes 20-50 seconds until
sanlock is initialized on all hosts, it takes few more seconds
until VDSM reports the storage domain status using repoStat to the engine.

VM with a lease cannot run if the status of the storage domain that holds
the VM lease is - {"acquired": false} according to the VDSM
repoStat.
Therefore, only when the VDSM reports {"acquired": true} for
the lease storage domain the VM will be able to run.

This patch adds a new filter policy unit that filters out all the hosts that
sanlock is not initialized on them yet.

Change-Id: Idd82d990b568b975117863a28aa4afd465491ae6
Bug-Url: https://bugzilla.redhat.com/1599732
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>