core: warn inter-mac-pool range overlap
Log and audit-log a warning when there are overlapping
mac ranges between mac pools. This warning will be logged
whenever the mac pool cache in engine is regenerated e.g.
on startup.

Overlapping ranges among mac pools cause engine to try
and assign the same mac address in two pools whenever the
mac pool cache in engine is regenerated. This causes the
cache to fail its own validation for uniqueness of each
address. The cache then becomes incosistent which makes
engine unstable.
This is a backport from master.

Change-Id: I9e45bcfb6c6d017493d8b2efc5a1eb15a7d0924f
Bug-Url: https://bugzilla.redhat.com/1639460
Signed-off-by: Eitan Raviv <eraviv@redhat.com>