core: set device alias during disk hot-plug
Since 4.2, we provide only device alias to VDSM when hot-unplugging
a device. If we don't set the device alias during the hot-plug
operation, there might be some time frame in which the device could
not be unplugged (since its alias remains empty until the next
iteration of VM devices monitoring).

Therefore, this patch changes hot-plug disk operations to also set
the device alias (in 4.2 and above, where user-aliases are used).
This was done for hot-plugging a NIC already.

Change-Id: Ia24467fef3a42c0debcba4ee3244267a7600c79a
Bug-Url: https://bugzilla.redhat.com/1613282
Signed-off-by: Arik Hadas <ahadas@redhat.com>