webadmin: allow copy of preallocated file-based images
The AsyncDataProvider#getvolumetype method determining whether
an image will be SPARSE or PREALLOCATED according to the image
format and the storage domain that the image is based on.

Since patch - 8140729f806e3efe81f92ef0355d5e84fd9403a6
VDSM is now supporting in keeping volume preallocation
during a copy of the VM disks in order to create the VM template
disks.

This patch changes the current policy that prevents from creating
preallocated file-based images via the UI when creating a template VM,
only if the cluster version > 4.2.

Change-Id: Ie2c7a96dc94832db08ec259d39028e0b59fc7649
Bug-Url: https://bugzilla.redhat.com/1615287
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>