core: UpdateVmCommand uses CompensationContext
Changes the code, so the command can optionally use
CompensationContext to store compensations for all
its DB operations.

This is not useful if the command is run independently,
because it still uses transaction.

It will be useful, in case this command is run as a child
of another command, and the parent fails and wants to
revert the effects of this command.

When this command is run as a child of UpdateClusterCommand,
there are no async tasks, because VM lease does not change
and neither does template version.

Change-Id: Ibe60e4c64ee400edd2e5683340f805f49ea6a978
Bug-Url: https://bugzilla.redhat.com/1609319
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>