core: UpdateVmTemplateCommand uses CompensationContext
Changes the code, so the command can optionally use
CompensationContext to store compensations for all
its DB operations.

This is not useful if the command is run independently,
because it still uses transaction.

It will be useful, in case this command is run as a child
of another command, and the parent fails and wants to
revert the effects of this command.

Change-Id: I702f51848138d9212b071927d96000bb707191e6
Bug-Url: https://bugzilla.redhat.com/1609319
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>