packaging: dbscripts: upgrade: Add constraint for auth_url
The authentication URL of provider has to match an expected
format, to ensure that the URL can be split to its components.

If there is an unexpected authentication URL in the db during
upgrade, engine-setup will interactively ask for an updated
value.

Change-Id: I23117079a88013527661bfde45b4da65f716dcf3
Bug-Url: https://bugzilla.redhat.com/1598391
Signed-off-by: Dominik Holler <dholler@redhat.com>
(cherry picked from commit 966861b0db034481caac9264d436cb97deda8585)