core: detach DestroyImageCommand from parent
When snapshot creation fails in the context of
Live Storage Migration, we might receive an NPE
due to a race between the endWithFailure phase of
LiveMigrateDiskCommand and the call to endAction of
CreateSnapshot command. If the LiveMigrateDiskCommand
executes first, CreateSnapshot will be cleared and
the attempt to execute endAction will result in NPE.

This patch detaches the call to DestroyImageCommand
from CreateSnapshot, since it's a negative flow and
is unrelated to its execution. Thus, when the deleteImage
vdsm task is completed, the endAction of CreateSnapshot
will not be executed.

Change-Id: Ie00454c2111fb394c700ec2796832e3c9fb2dd24
Bug-Url: https://bugzilla.redhat.com/1598996
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>