host-deploy: use a blocking factor of 1 for tar
The block size and record size that TarArchiveOutputStream use to create
the ovirt-host-deploy tarball have changed to be 512 bytes.
This causes host-deploy to hang on tar extraction on fc28.  Since the
block and record sizes cannot be controlled from TarArchiveOutputStream,
we need to ask tar to use a blocking factor of 1 (meaning 1x512 bytes).

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1460625
Change-Id: I2d84c2bcda73166d3ece38448b18f35f448db2ea
Signed-off-by: Yuval Turgeman <yturgema@redhat.com>