webadmin: Show SATA in the disk interfaces list
Show SATA in the disk interfaces dropdown list when adding/editing VM's
disks. For VMs using q35 chipset, IDE is hidden from the list and SATA
is shown instead. To achieve this, osinfo diskInterfaces property syntax
was changed - now it is possible to prefix some interface names with
chipset name. These particular interfaces will be available only for
this particular chipset. For example:

  [...].diskInterfaces.value = i440fx/IDE, VirtIO, q35/SATA

Change-Id: I907f2002bf23de490a15f570186cacb30a24327a
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>