core: Add Validation of AuthUrl to ProviderValidator
The authentication URL of provider has to match an expected
format, to ensure that the URL can be slitted in it's
components.

This change extends the ProviderValidator to check if the
AuthUrl is in the expected format during Adding or Updating a
provider.

Change-Id: I3ec84020dd7e8da219a9fadde5b076f3791a516e
Bug-Url: https://bugzilla.redhat.com/1598391
Signed-off-by: Dominik Holler <dholler@redhat.com>
(cherry picked from commit a0843209f6a28a42aae6f0d3606817e1920e9485)