test, do not merge: Revert "host-deploy: use a blocking factor of 1 for tar"
This reverts commit 1a4a8d617bbfb4202069d8a6bc33513694feb81b.

I suspect that that patch broke host-deploy, causing excess null bytes
to be sent to otopi - the rest of the padding in the tar file, which was
not consumed by tar following that patch.

If this actually fixes stuff, we need to find a way to fix the tar file,
not merge current revert patch.

Change-Id: I7b7aef2f02eb95760637b7691f3200e56a61513b
Signed-off-by: Yedidyah Bar David <didi@redhat.com>