webadmin: Use CSS to format the title label
The explicit calculation of the width of the title label of
networks in the HostSetupNetworksPopupView is replaced by a
CSS definition.

The explicit calculation was built on the assumption that
changing the width of the label would result in a change in
the size of the containing elements, which was not true for
for IE11 and Edge 42 and resulting in an endless loop for this
two browsers.

Change-Id: Ib10780b289d07bc74e692d1b007d39d694e5bcab
Bug-Url: https://bugzilla.redhat.com/1622321
Signed-off-by: Dominik Holler <dholler@redhat.com>