webadmin: handle empty vNIC profile in snapshots
If the vNIC profile is empty, the name of the vNIC profile and
the network are null, which is no valid input for
SafeHtmlUtils.fromString().
For this reason suitable strings are shown for the network name
and the emptyProfile name if the vNIC profile is empty.

Change-Id: I243a4d310315fc017a66427f1dae1ec041d7460e
Bug-Url: https://bugzilla.redhat.com/1585840
Signed-off-by: Dominik Holler <dholler@redhat.com>
(cherry picked from commit f7bcf66d740859be3458c73dc7fa595e9660d490)