core: unlock snapshot at the end of the command
Currently, when issuing a live snapshot, the snapshot
entity is unlock before the live snapshot is issued.
This may cause tools and operation relying on the status
of the entity to fail because the snapshot operation is not complete.

This patch moves the unlocking of the snapshot entity to the "endAction"
phase.

Change-Id: Ic9f6003c5f071f25a5136ae4c2cbe6d44ae7f861
Bug-Url: https://bugzilla.redhat.com/1628909
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>