packaging: find default MAC Pool by property
The default MAC Pool to add the default range is searched by
the dedicated property 'default_pool' instead of the UUID,
because this is more robust across different releases.

The stored SQL procedure GetDefaultMacPool is not used,
because its return value is not suitable to be used in
python.

Change-Id: I9468ff44ca6f07834f14ae2066579093e230696f
Bug-Url: https://bugzilla.redhat.com/1516445
Signed-off-by: Dominik Holler <dholler@redhat.com>
(cherry picked from commit 9fde48c64ca9d085dfb116439cdae046baf9f516)