webadmin: upload image - use actual size for transfer size
As part of the fix in commit fce2ae493dbf2fd75537d46224adb594b6b2c3df,
imageInfoModel.getActualSize() is used instead of 'getImageSize' method.

This patch ammends 'setTransferSize' and 'getImageInfoModel().validate'
methods to use imageInfoModel.getActualSize() as well. This is needed
as the transfer and validation size should be the image actual size.

Change-Id: I19c3769a8fca1c1e995473c82eba7ce038233701
Bug-Url: https://bugzilla.redhat.com/1627032
Signed-off-by: Daniel Erez <derez@redhat.com>