scheduler: Balancing tries to migrate multiple VMs until successful
Previously, the balancer tried to migrate only one VM,
that was considered a good candidate for balancing.
If the VM could not be migrated, the balancer would try
it again the next iteration anyway.

This patch changes the behavior. All candidate VMs
are ordered based on how suitable for balancing they are.
And then the balancer tries to migrate them in order, until one
migration succeeds.

Change-Id: Ie4b366a7852c1f734333761010111eefcd057167
Bug-Url: https://bugzilla.redhat.com/1583009
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>