core: fix failing 3.5 snapshot deletion
The engine fails to delete a snapshot that was created
in 3.5 cluster and imported/updated to a 4.2 cluster, it fails
since the engine is trying to update the VM's OVF and fetch
the config values from the DB according to 3.5 cluster level.

This is a regression that caused by
commit - 8c7b873a8afb93c4695293cfa86d97a8ff14a711
This commit avoids the use of clusterUtils.getCompatibilityVersion()
in order to prevent an override of the snapshot compatibility version
but allows receiving a compatibility version which is less than the
lowest compatibility version that the engine supports (3.6).

This patch sets the compatibility version to be the lowest compatibility
version that the engine allows if the compatibility version that received
from the snapshot static data is lower than it.

In this case, if the top image snapshot also created in a cluster
version < 3.6, the memory of the snapshot is not supported by the
engine and should be removed to prevent a preview of unsupported
snapshot memory.

Change-Id: Ie4dc4371bdc7f34080d209159d1d14b7728800fa
Bug-Url: https://bugzilla.redhat.com/1628150
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>